chore: Move date.h into subdirectory
