+++ src/utils/slot.cpp
@@ -3,7 +3,7 @@
 #include "utils/slot.h"
 
 #undef round
-#include "date.h"
+#include "date/date.h"
 
 #include <sstream>
 #include <cmath>
